Add license file
The repo did not contain a license file which, in theory, prevents the use of the license. See the following page for details: https://choosealicense.com/no-permission/

I've added the Apache license because that's what the GreenDroid project used which, according to the author, served as an inspiration for the author.